(exit)の修正、norm修正
