+++ src/exec/exec_cmd/builtin_cmd/exec_cmd_bi_cd.c
@@ -6,7 +6,7 @@
 /*   By: tkuwahat <tkuwahat@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/01 20:51:34 by tkuwahat          #+#    #+#             */
-/*   Updated: 2025/10/16 18:59:08 by tkuwahat         ###   ########.fr       */
+/*   Updated: 2025/10/17 00:36:40 by tkuwahat         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -62,7 +62,7 @@ int	bi_cd(char **av)
 	char		cwd[PATH_MAX];
 	const char	*dst;
 	int			print_after;
-	
+
 	if (has_too_many_args(av))
 		return (err3("minishell: cd: ", NULL, "too many arguments\n"), 1);
 	dst = NULL;
