+++ src/exec/exec_cmd/child/exec_cmd_child_main.c
@@ -6,7 +6,7 @@
 /*   By: tkuwahat <tkuwahat@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/01 20:04:49 by tkuwahat          #+#    #+#             */
-/*   Updated: 2025/10/16 18:53:27 by tkuwahat         ###   ########.fr       */
+/*   Updated: 2025/10/17 00:20:01 by tkuwahat         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
