+++ src/exec/exec_cmd/exec_cmd_single.c
@@ -6,7 +6,7 @@
 /*   By: tkuwahat <tkuwahat@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/29 20:33:07 by tkuwahat          #+#    #+#             */
-/*   Updated: 2025/10/16 18:41:04 by tkuwahat         ###   ########.fr       */
+/*   Updated: 2025/10/17 00:42:49 by tkuwahat         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -58,7 +58,7 @@ int	exec_single(t_cmd *c, t_exec_ctx *ctx)
 	if (!c)
 		return (set_exit_status(2), -1);
 	if (!in_sub_or_pipe_ctx(ctx) && is_builtin_parent(c))
-		return (run_parent_builtin_flow(c));
+		return (run_parent_flow(c, ctx));
 	if (is_no_argv_but_has_redirs(c))
 		return (redirect_only_flow(c));
 	if (is_noop_cmd(c))
